requirements.txt
streamlit
pandas
PyPDF2
python-docx
spacy
scikit-learn
nltk
reportlab